chore: move memoized test into separated file
create integratio test workflow
